IOS-2310 Migrate EditorSetRouter to EditorSetCoordinator
