+++ Anytype/Sources/PresentationLayer/Flows/EditorSetFlow/EditorSetCoordinatorView.swift
@@ -15,6 +15,9 @@ struct EditorSetCoordinatorView: View {
             .onChange(of: model.dismiss) { _ in
                 dismiss()
             }
+            .sheet(item: $model.setQueryData) { data in
+                model.setQuery(data)
+            }
             .anytypeSheet(item: $model.setViewPickerData) { data in
                 model.setViewPicker(data: data)
             }
