+++ Anytype/Sources/PresentationLayer/Flows/ObjectSettingsFlow/ObjectSettingsCoordinator.swift
@@ -2,6 +2,7 @@ import Foundation
 import Services
 import AnytypeCore
 
+@MainActor
 protocol ObjectSettingsCoordinatorProtocol {
     func startFlow(
         objectId: BlockId,
@@ -11,6 +12,7 @@ protocol ObjectSettingsCoordinatorProtocol {
     )
 }
 
+@MainActor
 final class ObjectSettingsCoordinator: ObjectSettingsCoordinatorProtocol,
                                        ObjectSettingsModelOutput, RelationsListModuleOutput,
                                        RelationValueCoordinatorOutput {
@@ -119,7 +121,7 @@ final class ObjectSettingsCoordinator: ObjectSettingsCoordinatorProtocol,
     }
     
     func openPageAction(screenData: EditorScreenData) {
-        output?.showPage(data: screenData)
+        output?.showEditorScreen(data: screenData)
     }
     
     func linkToAction(document: BaseDocumentProtocol, onSelect: @escaping (BlockId) -> ()) {
@@ -143,7 +145,6 @@ final class ObjectSettingsCoordinator: ObjectSettingsCoordinatorProtocol,
     
     // MARK: - RelationsListModuleOutput
     
-    @MainActor
     func addNewRelationAction(document: BaseDocumentProtocol) {
         addNewRelationCoordinator.showAddNewRelationView(
             document: document,
@@ -155,7 +156,6 @@ final class ObjectSettingsCoordinator: ObjectSettingsCoordinatorProtocol,
         )
     }
     
-    @MainActor
     func editRelationValueAction(document: BaseDocumentProtocol, relationKey: String) {
         let relation = document.parsedRelations.installed.first { $0.key == relationKey }
         guard let relation = relation else {
@@ -178,9 +178,9 @@ final class ObjectSettingsCoordinator: ObjectSettingsCoordinatorProtocol,
     
     // MARK: - RelationValueCoordinatorOutput
     
-    func openObject(screenData: EditorScreenData) {
+    func showEditorScreen(data: EditorScreenData) {
         navigationContext.dismissAllPresented(animated: true) { [weak self] in
-            self?.output?.showPage(data: screenData)
+            self?.output?.showEditorScreen(data: data)
         }
     }
 }
