+++ Anytype/Sources/PresentationLayer/LinkToObjectFlow/LinkToObjectCoordinatorOutput.swift
@@ -1,5 +1,5 @@
 import Foundation
 
 protocol LinkToObjectCoordinatorOutput: AnyObject {
-    func showPage(data: EditorScreenData)
+    func showEditorScreen(data: EditorScreenData)
 }
