+++ Anytype/Sources/PresentationLayer/TextEditor/Assembly/Set/EditorSetModuleAssembly.swift
@@ -69,44 +69,19 @@ final class EditorSetModuleAssembly: EditorSetModuleAssemblyProtocol {
             activeWorkspaceStorage: serviceLocator.activeWorkspaceStorage(),
             output: output
         )
-
-        let router = EditorSetRouter(
-            setDocument: setDocument,
-            navigationContext: uiHelpersDI.commonNavigationContext(),
-            createObjectModuleAssembly: modulesDI.createObject(),
-            newSearchModuleAssembly: modulesDI.newSearch(),
-            objectTypeSearchModuleAssembly: modulesDI.objectTypeSearch(),
-            objectSettingCoordinator: coordinatorsDI.objectSettings().make(),
-            relationValueCoordinator: coordinatorsDI.relationValue().make(), 
-            setObjectCreationCoordinator: coordinatorsDI.setObjectCreation().make(),
-            objectCoverPickerModuleAssembly: modulesDI.objectCoverPicker(),
-            objectIconPickerModuleAssembly: modulesDI.objectIconPicker(),
-            setViewSettingsCoordinatorAssembly: coordinatorsDI.setViewSettings(),
-            setSortsListCoordinatorAssembly: coordinatorsDI.setSortsList(),
-            setFiltersListCoordinatorAssembly: coordinatorsDI.setFiltersList(),
-            setViewSettingsImagePreviewModuleAssembly: modulesDI.setViewSettingsImagePreview(),
-            setViewSettingsGroupByModuleAssembly: modulesDI.setViewSettingsGroupByView(),
-            editorSetRelationsCoordinatorAssembly: coordinatorsDI.setRelations(),
-            setViewPickerCoordinatorAssembly: coordinatorsDI.setViewPicker(),
-            toastPresenter: uiHelpersDI.toastPresenter(),
-            setObjectCreationSettingsCoordinator: coordinatorsDI.setObjectCreationSettings().make(with: uiHelpersDI.commonNavigationContext()),
-            output: output
-        )
         
-        setupHeaderModelActions(headerModel: headerModel, using: router)
-        
-        model.setup(router: router)
+        setupHeaderModelActions(headerModel: headerModel, using: output)
         
         return model
     }
     
-    private func setupHeaderModelActions(headerModel: ObjectHeaderViewModel, using router: ObjectHeaderRouterProtocol) {
-        headerModel.onCoverPickerTap = { [weak router] args in
-            router?.showCoverPicker(document: args.0, onCoverAction: args.1)
+    private func setupHeaderModelActions(headerModel: ObjectHeaderViewModel, using output: EditorSetModuleOutput?) {
+        headerModel.onCoverPickerTap = { args in
+            output?.showCoverPicker(document: args.0, onCoverAction: args.1)
         }
         
-        headerModel.onIconPickerTap = { [weak router] args in
-            router?.showIconPicker(document: args.0, onIconAction: args.1)
+        headerModel.onIconPickerTap = { args in
+            output?.showIconPicker(document: args.0, onIconAction: args.1)
         }
     }
 }
