+++ Anytype/Sources/PresentationLayer/TextEditor/Assembly/Set/EditorSetModuleOutput.swift
@@ -5,8 +5,34 @@ import Services
 protocol EditorSetModuleOutput: AnyObject {
     func showEditorScreen(data: EditorScreenData)
     func replaceEditorScreen(data: EditorScreenData)
+    func closeEditor()
+    
     func showSetViewPicker(document: SetDocumentProtocol, subscriptionDetailsStorage: ObjectDetailsStorage)
     func showSetViewSettings(document: SetDocumentProtocol, subscriptionDetailsStorage: ObjectDetailsStorage)
-    func closeEditor()
-    // Migrate EditorSetRouter to EditorSetCoordinator
+    func showQueries(document: SetDocumentProtocol, selectedObjectId: BlockId?, onSelect: @escaping (BlockId) -> ())
+
+    // NavigationContext
+    func showCreateObject(document: SetDocumentProtocol, setting: ObjectCreationSetting?)
+    func showKanbanColumnSettings(
+        hideColumn: Bool,
+        selectedColor: BlockBackgroundColor?,
+        onSelect: @escaping (Bool, BlockBackgroundColor?) -> Void
+    )
+    func showSettings(actionHandler: @escaping (ObjectSettingsAction) -> Void)
+    func showCoverPicker(
+        document: BaseDocumentGeneralProtocol,
+        onCoverAction: @escaping (ObjectCoverPickerAction) -> Void
+    )
+    func showIconPicker(
+        document: BaseDocumentGeneralProtocol,
+        onIconAction: @escaping (ObjectIconPickerAction) -> Void
+    )
+    func showRelationValueEditingView(document: SetDocumentProtocol, key: String)
+    func showRelationValueEditingView(objectDetails: ObjectDetails, relation: Relation)
+    func showSetObjectCreationSettings(
+        document: SetDocumentProtocol,
+        viewId: String,
+        onTemplateSelection: @escaping (ObjectCreationSetting) -> ()
+    )
+    func showFailureToast(message: String)
 }
