+++ Anytype/Sources/PresentationLayer/TextEditor/EditorPage/Models/BlockViewModelBuilder.swift
@@ -218,7 +218,7 @@ final class BlockViewModelBuilder {
                 details: details,
                 detailsService: detailsService,
                 openLink: { [weak self] data in
-                    self?.router.showPage(data: data)
+                    self?.router.showEditorScreen(data: data)
                 }
             )
         case .featuredRelations:
@@ -297,7 +297,7 @@ final class BlockViewModelBuilder {
                     self?.router.showFailureToast(message: message)
                 },
                 openSet: { [weak self] data in
-                    self?.router.showPage(data: data)
+                    self?.router.showEditorScreen(data: data)
                 }
             )
         case .unsupported:
