+++ Anytype/Sources/PresentationLayer/TextEditor/EditorPage/Views/Settings/Relations/RelationsListModuleAsembly.swift
@@ -3,14 +3,14 @@ import UIKit
 import Services
 
 protocol RelationsListModuleAssemblyProtocol {
- 
+    @MainActor
     func make(document: BaseDocumentProtocol, output: RelationsListModuleOutput) -> UIViewController
 }
 
 final class RelationsListModuleAssembly: RelationsListModuleAssemblyProtocol {
     
     // MARK: - RelationsListModuleAssemblyProtocol
-    
+    @MainActor
     func make(document: BaseDocumentProtocol, output: RelationsListModuleOutput) -> UIViewController {
         
         let viewModel = RelationsListViewModel(
