+++ Anytype/Sources/PresentationLayer/TextEditor/EditorPage/Views/Settings/Relations/RelationsListModuleOutput.swift
@@ -1,5 +1,6 @@
 import Foundation
 
+@MainActor
 protocol RelationsListModuleOutput: AnyObject {
     func addNewRelationAction(document: BaseDocumentProtocol)
     func editRelationValueAction(document: BaseDocumentProtocol, relationKey: String)
