+++ Anytype/Sources/PresentationLayer/TextEditor/Set/EditorSetViewModel.swift
@@ -123,8 +123,7 @@ final class EditorSetViewModel: ObservableObject {
     
     let setDocument: SetDocumentProtocol
     let paginationHelper = EditorSetPaginationHelper()
-    
-    private var router: EditorSetRouterProtocol?
+
     private let subscriptionStorageProvider: SubscriptionStorageProviderProtocol
     private let dataviewService: DataviewServiceProtocol
     private let searchService: SearchServiceProtocol
@@ -171,11 +170,10 @@ final class EditorSetViewModel: ObservableObject {
         self.activeWorkspaceStorage = activeWorkspaceStorage
         self.output = output
         self.syncStatusData = SyncStatusData(status: .unknown, networkId: activeWorkspaceStorage.workspaceInfo.networkId)
+        self.setup()
     }
     
-    func setup(router: EditorSetRouterProtocol) {
-        self.router = router
-        
+    private func setup() {
         setDocument.setUpdatePublisher.sink { [weak self] update in
             Task { [weak self] in
                 await self?.onDataChange(update)
@@ -222,7 +220,7 @@ final class EditorSetViewModel: ObservableObject {
 
     func onRelationTap(relation: Relation) {
         if relation.hasSelectedObjectsRelationType {
-            router?.showFailureToast(message: Loc.Set.SourceType.Cancel.Toast.title)
+            output?.showFailureToast(message: Loc.Set.SourceType.Cancel.Toast.title)
         } else {
             showRelationValueEditingView(key: relation.key)
         }
@@ -543,8 +541,8 @@ final class EditorSetViewModel: ObservableObject {
     
     @MainActor
     func onSecondaryCreateTap() {
-        router?.showSetObjectCreationSettings(
-            setDocument: setDocument,
+        output?.showSetObjectCreationSettings(
+            document: setDocument,
             viewId: activeView.id,
             onTemplateSelection: { [weak self] setting in
                 self?.createObject(setting: setting)
@@ -561,7 +559,7 @@ final class EditorSetViewModel: ObservableObject {
     }
     
     private func createObject(setting: ObjectCreationSetting? = nil) {
-        router?.showCreateObject(setting: setting)
+        output?.showCreateObject(document: setDocument, setting: setting)
     }
 
     private func defaultSubscriptionDetailsStorage(file: StaticString = #file, function: String = #function, line: UInt = #line) -> ObjectDetailsStorage? {
@@ -582,7 +580,7 @@ extension EditorSetViewModel {
             return
         }
 
-        router?.showRelationValueEditingView(key: key)
+        output?.showRelationValueEditingView(document: setDocument, key: key)
     }
     
     func showRelationValueEditingView(
@@ -595,7 +593,7 @@ extension EditorSetViewModel {
             return
         }
         
-        router?.showRelationValueEditingView(
+        output?.showRelationValueEditingView(
             objectDetails: objectDetails,
             relation: relation
         )
@@ -612,7 +610,7 @@ extension EditorSetViewModel {
     }
     
     func showObjectSettings() {
-        router?.showSettings { [weak self] action in
+        output?.showSettings { [weak self] action in
             switch action {
             case .cover(let objectCoverPickerAction):
                 self?.headerModel.handleCoverAction(action: objectCoverPickerAction)
@@ -637,7 +635,7 @@ extension EditorSetViewModel {
         let groupOrder = setDocument.dataView.groupOrders.first { [weak self] in $0.viewID == self?.activeView.id }
         let viewGroup = groupOrder?.viewGroups.first { $0.groupID == groupId }
         let selectedColor = MiddlewareColor(rawValue: viewGroup?.backgroundColor ?? "")?.backgroundColor
-        router?.showKanbanColumnSettings(
+        output?.showKanbanColumnSettings(
             hideColumn: viewGroup?.hidden ?? false,
             selectedColor: selectedColor,
             onSelect: { [weak self] hidden, backgroundColor in
@@ -651,13 +649,13 @@ extension EditorSetViewModel {
     }
     
     func showIconPicker() {
-        router?.showIconPicker(document: setDocument) { [weak self] action in
+        output?.showIconPicker(document: setDocument) { [weak self] action in
             self?.headerModel.handleIconAction(action: action)
         }
     }
     
     func showSetOfTypeSelection() {
-        router?.showQueries(selectedObjectId: setDocument.details?.setOf.first) { [weak self] typeObjectId in
+        output?.showQueries(document: setDocument, selectedObjectId: setDocument.details?.setOf.first) { [weak self] typeObjectId in
             guard let self else { return }
             Task { @MainActor in
                 try? await self.objectActionsService.setSource(objectId: self.objectId, source: [typeObjectId])
